AlexNet implementation only works for 224x224
Tensor flattening of torchvision.models.AlexNet is not agnostic to input shape. The documentation says it should work as long as input image bigger than 224x224. The code does not comply with this. 

I propose to change the x.view sentence to be x = x.view(x.size(0), -1).

Regards
Mo